Fix issues with Drizzle not building the migration due to decorators. Moved enums out of type-graphql files to resolve this issues.
Fixed file name for RunMigration.ts.

Updated Dockerfile to exclude drizzle migration generate (or deploy) command. This should be run separately.

Updated package.js to move drizzle-kit to dependencies and upgrade ts-node to latest released version.
