+++ package.json
@@ -41,9 +41,8 @@
     "@types/node": "^14.14.16",
     "@types/pg": "^8.10.9",
     "concurrently": "^5.3.0",
-    "drizzle-kit": "^0.20.17",
     "nodemon": "^2.0.6",
-    "ts-node": "^9.1.1",
+    "ts-node": "^10.9.2",
     "typescript": "^4.1.3"
   },
   "dependencies": {
@@ -58,6 +57,7 @@
     "dotenv": "^8.2.0",
     "dotenv-expand": "^8.0.3",
     "dotenv-safe": "^8.2.0",
+    "drizzle-kit": "^0.20.17",
     "drizzle-orm": "^0.30.9",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
