+++ src/api/entities/Media/Media.ts
@@ -1,25 +1,7 @@
 import { Field, Int, ObjectType, registerEnumType } from "type-graphql";
 import { Post } from "../Post";
 import { User } from "../User";
-
-export enum MediaTypes {
-  Audio = "AUDIO",
-  Model = "MODEL",
-  Photo = "PHOTO",
-  Video = "VIDEO",
-}
-
-export enum MediaExtensions {
-  PNG = "PNG",
-  JPG = "JPG",
-  JPEG = "JPEG",
-  BMP = "BMP",
-  MP4 = "MP4",
-  AVI = "AVI",
-  FBX = "FBX",
-  USDZ = "USDZ",
-  GLTF = "GLTF",
-}
+import { MediaTypes, MediaExtensions } from "../../enums";
 
 registerEnumType(MediaTypes, {
   name: "MediaTypes",
