+++ src/api/enums/index.ts
@@ -1,1 +1,62 @@
+export enum LocationTypes {
+  Place = "Place",
+  User = "User",
+}
 
+export enum MediaTypes {
+  Audio = "AUDIO",
+  Model = "MODEL",
+  Photo = "PHOTO",
+  Video = "VIDEO",
+}
+
+export enum MediaExtensions {
+  PNG = "PNG",
+  JPG = "JPG",
+  JPEG = "JPEG",
+  BMP = "BMP",
+  MP4 = "MP4",
+  AVI = "AVI",
+  FBX = "FBX",
+  USDZ = "USDZ",
+  GLTF = "GLTF",
+}
+
+export enum PlaceTypes {
+  Administrative = "Administrative",
+  Commercial = "Commercial",
+  Community = "Community",
+  Continent = "Continent",
+  Country = "Country",
+  Educational = "Educational",
+  Geographic = "Geographic",
+  Landmark = "Landmark",
+  Medical = "Medical",
+  Municipality = "Municipality",
+  Province = "Province",
+  Religious = "Religious",
+  Residential = "Residential",
+  State = "State",
+  Transit = "Transit",
+}
+
+export enum UserRoles {
+  Admin = "ADMIN",
+  User = "USER",
+}
+
+export enum UserStatuses {
+  Active = "ACTIVE",
+  Archive = "ARCHIVE",
+  Deactivated = "DEACTIVATED",
+  Inactive = "INACTIVE",
+  Invited = "INVITED",
+  Removed = "REMOVED",
+}
+
+export enum UserVerifications {
+  Celebrity = "CELEBRITY",
+  Developer = "DEVELOPER",
+  Official = "OFFICIAL",
+  Standard = "STANDARD",
+}
