+++ src/api/resolvers/PlaceResolver/inputs.ts
@@ -2,7 +2,7 @@ import { Field, Float, InputType, Int } from "type-graphql";
 import { Coordinates } from "../../entities/Coordinates/Coordinates";
 import { AddressInput } from "../AddressResolver";
 import { PlaceIncludes } from "./types";
-import { PlaceTypes } from "../../entities/Place";
+import { PlaceTypes } from "../../enums";
 
 @InputType()
 export class CreatePlaceInput {
