+++ src/api/resolvers/PostResolver/inputs.ts
@@ -2,7 +2,7 @@ import { Point } from "geojson";
 import { Field, InputType } from "type-graphql";
 import { MediaScalar } from "../../../types/MediaScalar";
 import { PointScalar } from "../../../types/Point";
-import { Media } from "../../entities/Media/Media";
+import { Media } from "../../entities/Media";
 
 @InputType()
 export class PostCreateInput {
