+++ src/api/schema/media.ts
@@ -13,7 +13,7 @@ import {
 } from "drizzle-orm/pg-core";
 import { relations, sql } from "drizzle-orm";
 import { posts } from "./posts";
-import { MediaTypes, MediaExtensions } from "../Media";
+import { MediaTypes, MediaExtensions } from "../enums";
 
 export const MediaTypesDZL = pgEnum(
   "media_types",
