+++ src/api/schema/posts.ts
