+++ src/lib/drizzle/index.ts
@@ -1,5 +1,5 @@
 import { drizzle } from "drizzle-orm/node-postgres";
 import { dbPool } from "../pg";
-import * as schema from "../../api/entities/Schema";
+import * as schema from "../../api/schema";
 
 export const dzlClient = drizzle(dbPool, { schema });
